requirements.txt
streamlit
Pandas
Pillow
google-generativeai
python-dotenv